# StarWars/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "StarWars.Api",
    "version": "1.0"
  },
  "paths": {
    "/Film": {
      "post": {
        "tags": [
          "Film"
        ],
        "operationId": "CreateFilm",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/FilmCreateForm"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmCreateForm"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmCreateForm"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FilmCreateForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FilmCreateResultViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmCreateResultViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmCreateResultViewModel"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Film"
        ],
        "operationId": "GetFilms",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FilmListViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmListViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmListViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/StarWars/Characters/{characterId}": {
      "get": {
        "tags": [
          "StarWarsCharacters"
        ],
        "operationId": "GetStarWarsCharacter",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterViewModel"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "StarWarsCharacters"
        ],
        "operationId": "EditStarWarsCharacter",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EditCharacterForm"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCharacterForm"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCharacterForm"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditCharacterForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "StarWarsCharacters"
        ],
        "operationId": "DeleteStarWarsCharacter",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/StarWars/Characters/Search": {
      "post": {
        "tags": [
          "StarWarsCharacters"
        ],
        "operationId": "SearchStarWarsCharacter",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchCharacterForm"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchCharacterForm"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchCharacterForm"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchCharacterForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCharacterResultViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCharacterResultViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCharacterResultViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/StarWars/Characters/Search/Data": {
      "get": {
        "tags": [
          "StarWarsCharacters"
        ],
        "operationId": "GetSearchStarWarsCharacterData",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCharacterDataViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCharacterDataViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCharacterDataViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/StarWars/Characters": {
      "post": {
        "tags": [
          "StarWarsCharacters"
        ],
        "operationId": "CreateStarWarsCharacter",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterForm"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterForm"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterForm"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCharacterResultViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCharacterResultViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCharacterResultViewModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CharacterViewModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "originalName": {
            "type": "string",
            "nullable": true
          },
          "birthYear": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "race": {
            "type": "string",
            "nullable": true
          },
          "height": {
            "type": "string",
            "nullable": true
          },
          "hairColor": {
            "type": "string",
            "nullable": true
          },
          "eyeColor": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "originPlanetName": {
            "type": "string",
            "nullable": true
          },
          "films": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateCharacterForm": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "originalName": {
            "type": "string",
            "nullable": true
          },
          "birthYearBBY": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "race": {
            "type": "string",
            "nullable": true
          },
          "heightMeters": {
            "type": "number",
            "format": "float"
          },
          "hairColor": {
            "type": "string",
            "nullable": true
          },
          "eyeColor": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "originPlanetName": {
            "type": "string",
            "nullable": true
          },
          "filmIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateCharacterResultViewModel": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "EditCharacterForm": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "originalName": {
            "type": "string",
            "nullable": true
          },
          "birthYearBBY": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "race": {
            "type": "string",
            "nullable": true
          },
          "heightMeters": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "hairColor": {
            "type": "string",
            "nullable": true
          },
          "eyeColor": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "originPlanetName": {
            "type": "string",
            "nullable": true
          },
          "filmIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FilmCreateForm": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "FilmCreateResultViewModel": {
        "type": "object",
        "properties": {
          "filmId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "FilmListViewModel": {
        "type": "object",
        "properties": {
          "films": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilmViewModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FilmViewModel": {
        "type": "object",
        "properties": {
          "filmId": {
            "type": "string",
            "format": "uuid"
          },
          "filmName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Gender": {
        "enum": [
          "Male",
          "Female"
        ],
        "type": "string"
      },
      "GenderSelectItem": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "$ref": "#/components/schemas/Gender"
          }
        },
        "additionalProperties": false
      },
      "GuidSelectItem": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "SearchCharacterDataViewModel": {
        "type": "object",
        "properties": {
          "films": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuidSelectItem"
            },
            "nullable": true
          },
          "planets": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "genders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenderSelectItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SearchCharacterForm": {
        "type": "object",
        "properties": {
          "fromBirthYearBBY": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "toBirthYearBBY": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "films": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "planetName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "pageNumber": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageSize": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SearchCharacterResultViewModel": {
        "type": "object",
        "properties": {
          "characters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchCharacterViewModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SearchCharacterViewModel": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "originName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}